+++ bin/ci.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
 
+set +x
+
 bin/bundle install --jobs 3 --retry 3 --path .bundle/gems
 bin/rake
+
+set -x