+++ lib/byebug/remote.rb
@@ -2,11 +2,11 @@
 
 require "socket"
 require "byebug/processors/control_processor"
+require "byebug/remote/server"
+require "byebug/remote/client"
 
 #
 # Remote debugging functionality.
-#
-# @todo Refactor & add tests
 #
 module Byebug
   # Port number used for remote debugging
@@ -17,8 +17,14 @@ module Byebug
     attr_accessor :wait_connection
 
     # The actual port that the server is started at
-    attr_accessor :actual_port
-    attr_reader :actual_control_port
+    def actual_port
+      server.actual_port
+    end
+
+    # The actual port that the control server is started at
+    def actual_control_port
+      control.actual_port
+    end
 
     #
     # Interrupts the current thread
@@ -28,81 +34,52 @@ module Byebug
     end
 
     #
-    # Starts a remote byebug
+    # Starts the remote server main thread
     #
     def start_server(host = nil, port = PORT)
-      return if @thread
-
-      Context.interface = nil
-      start
-
       start_control(host, port.zero? ? 0 : port + 1)
 
-      mutex = Mutex.new
-      proceed = ConditionVariable.new
-
-      server = TCPServer.new(host, port)
-      self.actual_port = server.addr[1]
-
-      yield if block_given?
-
-      @thread = DebugThread.new do
-        while (session = server.accept)
-          Context.interface = RemoteInterface.new(session)
-          mutex.synchronize { proceed.signal } if wait_connection
-        end
-      end
-
-      mutex.synchronize { proceed.wait(mutex) } if wait_connection
+      server.start(host, port)
     end
 
-    def start_control(host = nil, ctrl_port = PORT + 1)
-      return @actual_control_port if @control_thread
-      server = TCPServer.new(host, ctrl_port)
-      @actual_control_port = server.addr[1]
-
-      @control_thread = DebugThread.new do
-        while (session = server.accept)
-          context = Byebug.current_context
-          interface = RemoteInterface.new(session)
-
-          ControlProcessor.new(context, interface).process_commands
-        end
-      end
-
-      @actual_control_port
+    #
+    # Starts the remote server control thread
+    #
+    def start_control(host = nil, port = PORT + 1)
+      control.start(host, port)
     end
 
     #
     # Connects to the remote byebug
     #
     def start_client(host = "localhost", port = PORT)
-      interface = LocalInterface.new
-      puts "Connecting to byebug server at #{host}:#{port}..."
-      socket = TCPSocket.new(host, port)
-      puts "Connected."
-
-      while (line = socket.gets)
-        case line
-        when /^PROMPT (.*)$/
-          input = interface.read_command(Regexp.last_match[1])
-          break unless input
-          socket.puts input
-        when /^CONFIRM (.*)$/
-          input = interface.readline(Regexp.last_match[1])
-          break unless input
-          socket.puts input
-        else
-          puts line
-        end
-      end
-
-      socket.close
+      client.start(host, port)
     end
 
     def parse_host_and_port(host_port_spec)
       location = host_port_spec.split(":")
       location[1] ? [location[0], location[1].to_i] : ["localhost", location[0]]
+    end
+
+    private
+
+    def client
+      @client ||= Remote::Client.new(Context.interface)
+    end
+
+    def server
+      @server ||= Remote::Server.new(wait_connection: wait_connection) do |s|
+        Context.interface = RemoteInterface.new(s)
+      end
+    end
+
+    def control
+      @control ||= Remote::Server.new(wait_connection: false) do |s|
+        context = Byebug.current_context
+        interface = RemoteInterface.new(s)
+
+        ControlProcessor.new(context, interface).process_commands
+      end
     end
   end
 end